% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blouchOU.setup.v1.R,
%   R/blouchOUPredict.setup.v1.R
\name{sigma.X.estimate}
\alias{sigma.X.estimate}
\title{This is the R setup file for blouchOU_v1.stan
It includes the function sigma.X.estimate, taken from the R Pacakge Slouch (Kopperud et al. 2020).}
\usage{
\method{sigma}{X.estimate}(phy, ta, predictor, mv.predictor)

\method{sigma}{X.estimate}(phy, ta, predictor, mv.predictor)
}
\arguments{
\item{trdata}{Data formatted by make.treedata function from the R Pacakge treeplyr - only extant species}

\item{trdata.fos}{Data formatted by make.treedata function from the R Pacakge treeplyr - extant and fossil}

\item{names.traits}{Vector of trait names}

\item{classical}{Numeric denoting whether classical regression (=1) or inverse regression (=0) should be used in prediction}
}
\value{
An object of class list for use in blouchOU_v1.stan

An object of class list for use in blouchOU_v1.stan
}
\description{
This is the R setup file for blouchOU_v1.stan
It includes the function sigma.X.estimate, taken from the R Pacakge Slouch (Kopperud et al. 2020).

This is the R setup file for blouchOUPredict_v1.stan
It includes the function sigma.X.estimate, taken from the R Pacakge Slouch (Kopperud et al. 2020).
}
